- add anim to core libs - update convert dp to px, change param type from int to float

+++ TVCHANNELS/tnpro.core/src/main/java/com/tnpro/core/utils/ViewUtils.java
@@ -4,11 +4,11 @@ import android.content.Context;
 import android.util.TypedValue;
 
 public class ViewUtils {
-    public static int dpToPx(Context context, int dp) {
+    public static int dpToPx(Context context, float dp) {
         return (int)(dp * context.getResources().getDisplayMetrics().density);
     }
 
-    public static int spToPx(Context context, int sp) {
+    public static int spToPx(Context context, float sp) {
         return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, sp, context.getResources().getDisplayMetrics());
     }
 }
